refactor(workflows): extract work/work-pr definitions to auto-discovered modules
Extract workflow step definitions and artifact rules from enforce-step-workflow.js into separate workflow-definition.js files (work/ and work-pr/) to follow Open/Closed Principle — add new workflows by dropping a definition file, not modifying the orchestrator. Add discoverWorkflows() to load definitions dynamically. Update agent instructions with explicit ${TASKS_BASE}/${TICKET_ID}/* paths for planning artifacts (brief.md, spec.md, tasks.md).
